% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pane.R
\name{plot_pane}
\alias{plot_pane}
\title{plot_pane}
\usage{
plot_pane(pane_result, output_name, type_ggplot = NULL)
}
\arguments{
\item{pane_result}{The matrix returned by pane()}

\item{output_name}{output name for pane pdf plot}

\item{type_ggplot}{NULL by default or 'YES' to plot with ggplot2}
}
\value{
Returns an pane_plot.pdf file in the working directory
}
\description{
plot_pane allows to plot pane assignations as barplot, using either R base (default) or ggplot2 (type_ggplot = 'YES')
}
