% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{write_pane}
\alias{write_pane}
\title{write_pane}
\usage{
write_pane(pane_input, output_name)
}
\arguments{
\item{pane_input}{R list returned by pane() function}

\item{output_name}{string containing the file output name}
}
\description{
write_pane allows to save pane results in a table-like format.
}
\examples{
\dontrun{
pca = read_eigen(pca_input = 'data/TOY.pca.evec')
example_as = read.table('data/Example_AS', header=TRUE)
pane_results <- pane(pca_input = pca, as_file = example_as)
write_pane(pane_results, output_name = 'my_dir/my_pane_results.txt')
}
}
